+++ tools/cmd/seidb/benchmark_write.go
@@ -5,8 +5,8 @@ import (
 	"io/fs"
 	"os"
 
-	"github.com/sei-protocol/sei-db/benchmark/dbbackend"
 	"github.com/sei-protocol/sei-db/ss"
+	"github.com/sei-protocol/sei-db/tools/dbbackend"
 	"github.com/spf13/cobra"
 )
